feat: Add support for modsecurity secruleengine configuration
In support of https://gitlab.com/gitlab-org/gitlab/issues/8558, this
MR adds support for passing `MODSECURITY_SEC_RULE_ENGINE` to the
underlying chart to customize execution of modsecurity within the
ingress nginx configuration.

This configuration is enabled as "DetectionOnly" by default within the
nginx-controller ConfigMap but overridable per ingress resource. This
allows customization of the specific handling of the ruleset, in this
case enabling the engine to be toggled to "Off" or "On" to either
disable rulechecks entirely or enable blocking mode.

This MR is paired with
https://gitlab.com/gitlab-org/charts/auto-deploy-app/merge_requests/12
which adds the necessary fields to the chart