fix: Initialize DB on stable track
In https://gitlab.com/gitlab-org/gitlab/issues/30465 we found that
non-stable releasese, e.g. canary and incremental rollouts, were not
working as intended when DB_INITIALIZE was set.

This was due to `helm upgrade` "failing" if no resources will be
created, see https://github.com/helm/helm/issues/4670. This affected us
because in the DB_INITIALIZE case, we do not create any deployments nor
do we activate any services.

The solution proposed in this commit is to always run the database
initialization step on the stable track, with services and postgres
configured accordingly. The fact that initializeCommand is set should be
enough to prevent deployments (this is how the chart works), but for
additional safety we set replicas to 0.