Only run publish-dryrun on this repository
In a community contribution MR, https://gitlab.com/gitlab-org/cluster-integration/auto-deploy-image/merge_requests/41,
we found that semantic release's dry run will generally fail on forks,
as it requires write permissions to the source repository (the fork)
in the CI job, which requires additional configuration on the fork.
Until we have a way to disable this check, I propose we do not run
this job on forks.