ci: fix template tests not running after split
We recently split up the template tests, but they were not actually
running in CI due to to the script using `go test .` instead of the
recursive `go test ./...`. This also fixes some other path-related
problems, namely the relative paths for the fixtures, and that
`test_helpers.go` must be in the same directory as the tests to count as
the same package.