fix: add default ROLLOUT_RESOURCE_TYPE=deployment
Needing to specify `ROLLOUT_RESOURCE_TYPE` is an obscure implementation
detail that most user should not need to care about. We don't even have
built-in support for other types, so we should not force users to set
it.

This is especially relevant to users who include
`Jobs/Deploy.gitlab-ci.yml` instead of `Auto-DevOps.gitlab-ci.yml`.