fix: allow --minify to fail w/o current-context
`kubectl config view --minify` needs a current-context
so it can show the information minified. Minify basically means to
remove all data that is not relevant to the current-context.

Since we call `kubectl config view --minify` in the root scope of the
file, every parent process that uses this file to call one of its
functions will endup reexecuting this piece of code. So by allowing
minify to fail when the context is not setup, we allow the user to call
use_kube_context before calling other scripts, which would avoid
breaking the execution.